third_party/WebKit/LayoutTests/http/tests/devtools/console/console-format-es6-expected.txt

CONSOLE MESSAGE: line 11: [object Promise]
CONSOLE MESSAGE: line 12: [object Promise]
CONSOLE MESSAGE: line 11: Symbol()
CONSOLE MESSAGE: line 12: Symbol()
CONSOLE MESSAGE: line 11: Symbol(a)
CONSOLE MESSAGE: line 12: Symbol(a)
CONSOLE MESSAGE: line 11: [object Object]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 11: [object Map]
CONSOLE MESSAGE: line 12: [object Map]
CONSOLE MESSAGE: line 11: [object WeakMap]
CONSOLE MESSAGE: line 12: [object WeakMap]
CONSOLE MESSAGE: line 11: [object Set]
CONSOLE MESSAGE: line 12: [object Set]
CONSOLE MESSAGE: line 11: [object WeakSet]
CONSOLE MESSAGE: line 12: [object WeakSet]
CONSOLE MESSAGE: line 11: [object Map]
CONSOLE MESSAGE: line 12: [object Map]
CONSOLE MESSAGE: line 11: [object Map]
CONSOLE MESSAGE: line 12: [object Map]
CONSOLE MESSAGE: line 11: [object Set]
CONSOLE MESSAGE: line 12: [object Set]
CONSOLE MESSAGE: line 11: [object Set]
CONSOLE MESSAGE: line 12: [object Set]
CONSOLE MESSAGE: line 11: [object Map]
CONSOLE MESSAGE: line 12: [object Map]
CONSOLE MESSAGE: line 11: [object Generator]
CONSOLE MESSAGE: line 12: [object Generator]
Tests that console properly displays information about ES6 features.

console-format-es6.html:11 Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0}
console-format-es6.html:12 [Promise]
globals[0]
Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0}
console-format-es6.html:11 Symbol()
console-format-es6.html:12 [Symbol()]
globals[1]
Symbol()
console-format-es6.html:11 Symbol(a)
console-format-es6.html:12 [Symbol(a)]
globals[2]
Symbol(a)
console-format-es6.html:11 {a: Symbol(), Symbol(a): 2}
console-format-es6.html:12 [{…}]
globals[3]
{a: Symbol(), Symbol(a): 2}
console-format-es6.html:11 Map(1) {{…} => {…}}
console-format-es6.html:12 [Map(1)]
globals[4]
Map(1) {{…} => {…}}
console-format-es6.html:11 WeakMap {{…} => {…}}
console-format-es6.html:12 [WeakMap]
globals[5]
WeakMap {{…} => {…}}
console-format-es6.html:11 Set(1) {{…}}
console-format-es6.html:12 [Set(1)]
globals[6]
Set(1) {{…}}
console-format-es6.html:11 WeakSet {{…}}
console-format-es6.html:12 [WeakSet]
globals[7]
WeakSet {{…}}
console-format-es6.html:11 Map(1) {Map(0) => WeakMap}
console-format-es6.html:12 [Map(1)]
globals[8]
Map(1) {Map(0) => WeakMap}
console-format-es6.html:11 Map(1) {Map(1) => WeakMap}
console-format-es6.html:12 [Map(1)]
globals[9]
Map(1) {Map(1) => WeakMap}
console-format-es6.html:11 Set(1) {WeakSet}
console-format-es6.html:12 [Set(1)]
globals[10]
Set(1) {WeakSet}
console-format-es6.html:11 Set(1) {WeakSet}
console-format-es6.html:12 [Set(1)]
globals[11]
Set(1) {WeakSet}
console-format-es6.html:11 Map(6) {" from str " => " to str ", undefined => undefined, null => null, 42 => 42, {…} => {…}, …}
console-format-es6.html:12 [Map(6)]
globals[12]
Map(6) {" from str " => " to str ", undefined => undefined, null => null, 42 => 42, {…} => {…}, …}
console-format-es6.html:11 genFunction {[[GeneratorStatus]]: "suspended"}
console-format-es6.html:12 [genFunction]
globals[13]
genFunction {[[GeneratorStatus]]: "suspended"}
Expanded all messages
console-format-es6.html:11 Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0}
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: -0
console-format-es6.html:12 [Promise]
    0: Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0}
    length: 1
    __proto__: Array(0)
globals[0]
Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0}
    __proto__: Promise
    [[PromiseStatus]]: "rejected"
    [[PromiseValue]]: -0
console-format-es6.html:11 Symbol()
console-format-es6.html:12 [Symbol()]
    0: Symbol()
    length: 1
    __proto__: Array(0)
globals[1]
Symbol()
console-format-es6.html:11 Symbol(a)
console-format-es6.html:12 [Symbol(a)]
    0: Symbol(a)
    length: 1
    __proto__: Array(0)
globals[2]
Symbol(a)
console-format-es6.html:11 {a: Symbol(), Symbol(a): 2}
    a: Symbol()
    getter: (...)
    Symbol(a): 2
    get getter: ƒ getter()
    __proto__: Object
console-format-es6.html:12 [{…}]
    0: {a: Symbol(), Symbol(a): 2}
    length: 1
    __proto__: Array(0)
globals[3]
{a: Symbol(), Symbol(a): 2}
    a: Symbol()
    getter: (...)
    Symbol(a): 2
    get getter: ƒ getter()
    __proto__: Object
console-format-es6.html:11 Map(1) {{…} => {…}}
    size: (...)
    __proto__: Map
    [[Entries]]: Array(1)
        0: {Object => Object}
        length: 1
console-format-es6.html:12 [Map(1)]
    0: Map(1) {{…} => {…}}
    length: 1
    __proto__: Array(0)
globals[4]
Map(1) {{…} => {…}}
    size: (...)
    __proto__: Map
    [[Entries]]: Array(1)
        0: {Object => Object}
        length: 1
console-format-es6.html:11 WeakMap {{…} => {…}}
    __proto__: WeakMap
    [[Entries]]: Array(1)
        0: {Object => Object}
        length: 1
console-format-es6.html:12 [WeakMap]
    0: WeakMap {{…} => {…}}
    length: 1
    __proto__: Array(0)
globals[5]
WeakMap {{…} => {…}}
    __proto__: WeakMap
    [[Entries]]: Array(1)
        0: {Object => Object}
        length: 1
console-format-es6.html:11 Set(1) {{…}}
    size: (...)
    __proto__: Set
    [[Entries]]: Array(1)
        0: Object
        length: 1
console-format-es6.html:12 [Set(1)]
    0: Set(1) {{…}}
    length: 1
    __proto__: Array(0)
globals[6]
Set(1) {{…}}
    size: (...)
    __proto__: Set
    [[Entries]]: Array(1)
        0: Object
        length: 1
console-format-es6.html:11 WeakSet {{…}}
    __proto__: WeakSet
    [[Entries]]: Array(1)
        0: Object
        length: 1
console-format-es6.html:12 [WeakSet]
    0: WeakSet {{…}}
    length: 1
    __proto__: Array(0)
globals[7]
WeakSet {{…}}
    __proto__: WeakSet
    [[Entries]]: Array(1)
        0: Object
        length: 1
console-format-es6.html:11 Map(1) {Map(0) => WeakMap}
    size: (...)
    __proto__: Map
    [[Entries]]: Array(1)
        0: {Map(0) => WeakMap}
        length: 1
console-format-es6.html:12 [Map(1)]
    0: Map(1) {Map(0) => WeakMap}
    length: 1
    __proto__: Array(0)
globals[8]
Map(1) {Map(0) => WeakMap}
    size: (...)
    __proto__: Map
    [[Entries]]: Array(1)
        0: {Map(0) => WeakMap}
        length: 1
console-format-es6.html:11 Map(1) {Map(1) => WeakMap}
    size: (...)
    __proto__: Map
    [[Entries]]: Array(1)
        0: {Map(1) => WeakMap}
        length: 1
console-format-es6.html:12 [Map(1)]
    0: Map(1) {Map(1) => WeakMap}
    length: 1
    __proto__: Array(0)
globals[9]
Map(1) {Map(1) => WeakMap}
    size: (...)
    __proto__: Map
    [[Entries]]: Array(1)
        0: {Map(1) => WeakMap}
        length: 1
console-format-es6.html:11 Set(1) {WeakSet}
    size: (...)
    __proto__: Set
    [[Entries]]: Array(1)
        0: WeakSet
        length: 1
console-format-es6.html:12 [Set(1)]
    0: Set(1) {WeakSet}
    length: 1
    __proto__: Array(0)
globals[10]
Set(1) {WeakSet}
    size: (...)
    __proto__: Set
    [[Entries]]: Array(1)
        0: WeakSet
        length: 1
console-format-es6.html:11 Set(1) {WeakSet}
    size: (...)
    __proto__: Set
    [[Entries]]: Array(1)
        0: WeakSet
        length: 1
console-format-es6.html:12 [Set(1)]
    0: Set(1) {WeakSet}
    length: 1
    __proto__: Array(0)
globals[11]
Set(1) {WeakSet}
    size: (...)
    __proto__: Set
    [[Entries]]: Array(1)
        0: WeakSet
        length: 1
console-format-es6.html:11 Map(6) {" from str " => " to str ", undefined => undefined, null => null, 42 => 42, {…} => {…}, …}
    size: (...)
    __proto__: Map
    [[Entries]]: Array(6)
        0: {" from str " => " to str "}
        1: {undefined => undefined}
        2: {null => null}
        3: {42 => 42}
        4: {Object => Object}
        5: {Array(1) => Array(1)}
        length: 6
console-format-es6.html:12 [Map(6)]
    0: Map(6) {" from str " => " to str ", undefined => undefined, null => null, 42 => 42, {…} => {…}, …}
    length: 1
    __proto__: Array(0)
globals[12]
Map(6) {" from str " => " to str ", undefined => undefined, null => null, 42 => 42, {…} => {…}, …}
    size: (...)
    __proto__: Map
    [[Entries]]: Array(6)
        0: {" from str " => " to str "}
        1: {undefined => undefined}
        2: {null => null}
        3: {42 => 42}
        4: {Object => Object}
        5: {Array(1) => Array(1)}
        length: 6
console-format-es6.html:11 genFunction {[[GeneratorStatus]]: "suspended"}
    __proto__: Generator
    [[GeneratorStatus]]: "suspended"
    [[GeneratorFunction]]: ƒ* ()
    [[GeneratorReceiver]]: Window
    [[GeneratorLocation]]: console-format-es6.html:56
    [[Scopes]]: Scopes[2]
console-format-es6.html:12 [genFunction]
    0: genFunction {[[GeneratorStatus]]: "suspended"}
    length: 1
    __proto__: Array(0)
globals[13]
genFunction {[[GeneratorStatus]]: "suspended"}
    __proto__: Generator
    [[GeneratorStatus]]: "suspended"
    [[GeneratorFunction]]: ƒ* ()
    [[GeneratorReceiver]]: Window
    [[GeneratorLocation]]: console-format-es6.html:56
    [[Scopes]]: Scopes[2]
